requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Faker==37.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.42.0
networkx==3.5
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
streamlit==1.45.1
streamlit-ace==0.1.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
